services/transform/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
httpx
python-dateutil
asyncpg
python-dotenv